// frontend/src/lib/schema.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/counterpoint": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    post: operations["generateCounterpoint"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    CounterpointRequest: {
      cf: string[];
      key: string;
      cf_part_id: components["schemas"]["PartId"];
      part_id: components["schemas"]["PartId"];
      species: components["schemas"]["Species"];
    };
    Duration: {
      value: components["schemas"]["Fraction"];
    };
    Fraction: {
      /** Format: int32 */
      numerator: number;
      /** Format: int32 */
      denominator: number;
    };
    FullScore: {
      key: components["schemas"]["Key"];
      time_signature: components["schemas"]["TimeSignature"];
      body: components["schemas"]["Score"];
    };
    Key: {
      tonic: components["schemas"]["NoteName"];
      mode: components["schemas"]["Mode"];
    };
    Measure: {
      notes: components["schemas"]["Note"][];
    };
    /** @enum {string} */
    Mode: "Major" | "Minor";
    Note: {
      value: components["schemas"]["Pitch"] | components["schemas"]["Rest"];
      duration: components["schemas"]["Duration"];
      attribute: components["schemas"]["ScoreAttrs"];
    };
    NoteName: {
      /** Format: int32 */
      value: number;
    };
    Octave: {
      /** Format: int32 */
      value: number;
    };
    /** @enum {string} */
    PartId: "SOPRANO" | "ALTO" | "TENOR" | "BASS";
    Pitch: {
      octave: components["schemas"]["Octave"];
      note_name: components["schemas"]["NoteName"];
    };
    Rest: Record<string, never>;
    Score: {
      parts: {
        [key: string]: components["schemas"]["Measure"][];
      };
    };
    ScoreAttrs: {
      is_tied_start: boolean;
    };
    /** @enum {string} */
    Species: "first" | "second" | "third" | "fourth" | "fifth";
    TimeSignature: {
      /** Format: int32 */
      beats: number;
      beat_type: components["schemas"]["Duration"];
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
  generateCounterpoint: {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody: {
      content: {
        "application/json": components["schemas"]["CounterpointRequest"];
      };
    };
    responses: {
      /** @description The request has succeeded. */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": components["schemas"]["FullScore"];
        };
      };
    };
  };
}